remove duplicate conftest
